Add network validation script executed in the sagemaker_ui_post_startup script
**Description**

This change introduces the network validation script which tests if certain AWS services are reachable by making read only API calls with a set timeout. If the call exceeds the timeout, the script infers that it was caused by a bad network setup such as not having access to the internet/ VPC endpoint to make the call. API calls that resolve (succeed or fail) within the timeout are inferred as having the proper network setup.

AWS services for Compute Connections and Git are checked in this script. More specifically, the script lists the datazone connections to see which services need to be checked.

The unreachable services are aggregated and are displayed by writing to the post-startup-status.json, which displays the error notification in the IDE.

**Testing**

Tested in a SMUS portal containing internet, no internet, and no internet with VPC Endpoints to Datazone and s3.