Update network_validation script to flag failing calls due to endpoint URL as unreachable
**Description**

If a user has access to the internet, but does not have a VPC Endpoint attached to their network, API calls made within the network_validation script may fail due to "Could not connect to the endpoint URL". This change ensures these errors are caught and surfaced to the user.

**Testing**

Tested in a SMUS portal containing internet, no internet, no internet with VPC Endpoints to Datazone and s3 AND internet with only certain VPC endpoints.